FIX: Pagable 수정 완료

+++ loanService/src/main/kotlin/com/naverfinancial/loanservice/service/AccountService.kt
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page
 
 interface AccountService {
     // 통장 리스트 조회
-    fun selectAccounts(ndi: String?, status : AccountTypeStatus, limit : Int, offset : Int) : Page<Account>
+    fun selectAccounts(ndi: String?, status : AccountTypeStatus, limit : Int, offset : Long) : Page<Account>
 
     // 마이너스 통장 신청
     fun openAccount(ndi: String, userCreditRating: UserCreditRating): Account
